@@ -8,7 +8,7 @@
     >
       splitfair
     </a>
-    
+
     <span
       v-if="desktopOnly"
       :class="{ menuOpened }"
@@ -32,67 +32,29 @@
 </template>
 
 <script>
-import { mapMutations } from 'vuex';
-
-import App from './components/App.vue';
-import DownloadAppSection from './components/DownloadAppSection.vue';
-import AboutSection from './components/AboutSection.vue';
-
-export default {
-  name: 'ApplicationComponent',
-  components: {
-    App,
-    DownloadAppSection,
-    AboutSection,
-  },
-  data() {
-    return {
-      menuOpened: false,
-      desktopOnly: true,
-    };
-  },
-  mounted() {
-    const mql = window.matchMedia('(max-width: 600px)');
-    this.desktopOnly = !mql.matches;
-    mql.addListener(this.screenTest);
-  },
-  methods: {
-    ...mapMutations('App', [
-      'increment',
-    ]),
-    test() {
-      const test = 'test';
-      return test;
-    },
-    moveAppSideways(value) {
-      this.menuOpened = value;
-    },
-    screenTest(e) {
-      this.desktopOnly = !e.matches;
-    }
-  },
-};
-
+import options from './Application.js';
+export default options;
 </script>
 
-<style lang="scss">
-@import 'styles/variables';
-$media-query-mobile: only screen and (max-width: 600px);
+<style scoped>
+@import url("./styles/variables.css");
+
 * {
   margin: 0;
   padding: 0;
 }
 
 body {
-  background-color: $light-bg;
-
-  @media only screen and (max-width: 600px) {
-    background-color: $dark-bg;
+  background-color: var(--light-bg);
+}
+@media (max-width: 600px) {
+  body {
+    background-color: var(--dark-bg);
   }
 }
 
 .Application {
-  font-family: 'Avenir', Helvetica, Arial, sans-serif;
+  font-family: "Avenir", Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
@@ -102,43 +64,40 @@ body {
   height: 100vh;
   position: relative;
   cursor: url(assets/cursor-red.png), auto;
+}
 
-  .Application__logo {
-    cursor: pointer;
-  }
+.Application__logo {
+  cursor: pointer;
+}
 
-  .Application__link {
-    position: absolute;
-    text-decoration: none;
-    color: $black;
-    font-size: 28px;
-    z-index: 1;
-    transition: opacity 250ms ease;
+.Application__link {
+  position: absolute;
+  text-decoration: none;
+  color: var(--black);
+  font-size: 28px;
+  z-index: 1;
+  transition: opacity 250ms ease;
+}
+.Application__link.menuOpened {
+  opacity: 0;
+  pointer-events: none;
+}
+.Application__link.logo {
+  top: 4vw;
+  left: 5vw;
+}
 
-    &.menuOpened {
-      opacity: 0;
-      pointer-events: none;
-    }
-
-    &.logo {
-      top: 4vw;
-      left: 5vw;
-    }
-  }
-
-  .Application__version {
-    position: absolute;
-    bottom: 4vw;
-    left: 5vw;
-    color: $black;
-    font-size: 28px;
-    opacity: 0.4;
-    transition: opacity 250ms ease;
-
-    &.menuOpened {
-      opacity: 0;
-      pointer-events: none;
-    }
-  }
+.Application__version {
+  position: absolute;
+  bottom: 4vw;
+  left: 5vw;
+  color: var(--black);
+  font-size: 28px;
+  opacity: 0.4;
+  transition: opacity 250ms ease;
+}
+.Application__version.menuOpened {
+  opacity: 0;
+  pointer-events: none;
 }
 </style>
