@@ -1,91 +1,102 @@
 <template>
-  <div class="Numberpad">
+  <div class="NumberPad">
     <div class="button-row">
       <div
-        ref="button-7"
+        ref="button7"
         v-touch:tap="onTap"
         class="button"
       >
         7
       </div>
+
       <div
-        ref="button-8"
+        ref="button8"
         v-touch:tap="onTap"
         class="button"
       >
         8
       </div>
+
       <div
-        ref="button-9"
+        ref="button9"
         v-touch:tap="onTap"
         class="button"
       >
         9
       </div>
     </div>
+
     <div class="button-row">
       <div
-        ref="button-4"
+        ref="button4"
         v-touch:tap="onTap"
         class="button"
       >
         4
       </div>
+
       <div
-        ref="button-5"
+        ref="button5"
         v-touch:tap="onTap"
         class="button"
       >
         5
       </div>
+
       <div
-        ref="button-6"
+        ref="button6"
         v-touch:tap="onTap"
         class="button"
       >
         6
       </div>
     </div>
+
     <div class="button-row">
       <div
-        ref="button-1"
+        ref="button1"
         v-touch:tap="onTap"
         class="button"
       >
         1
       </div>
+
       <div
-        ref="button-2"
+        ref="button2"
         v-touch:tap="onTap"
         class="button"
       >
         2
       </div>
+
       <div
-        ref="button-3"
+        ref="button3"
         v-touch:tap="onTap"
         class="button"
       >
         3
       </div>
     </div>
+
     <div class="button-row">
       <div
-        ref="button-Backspace"
+        ref="buttonBackspace"
         v-touch:tap="onTap"
         class="button clear"
       >
         {{ backIcon }}
       </div>
+
       <div
-        ref="button-0"
+        ref="button0"
         v-touch:tap="onTap"
         class="button"
       >
         0
       </div>
+
       <div
-        ref="button-Enter"
+        ref="buttonEnter"
         v-touch:tap="onTap"
         class="button enter"
       >
@@ -95,112 +106,110 @@
   </div>
 </template>
 
-<script>
-export default {
-  name: 'NumberpadCP',
-  data: function() {
-    return {
-      backIcon: '<',
-      timeout: null,
-    };
-  },
-  mounted() {
-    document.addEventListener('keydown', this.onKeyDown);
-    document.addEventListener('keyup', this.onKeyup);
-  },
-  methods: {
-    onTap(e) {
-      const el = e.currentTarget;
-      if (el) {
-        clearTimeout(this.timeout);
-        el.classList.add('pressed');
-        this.timeout = setTimeout(() => {
-          el.classList.remove('pressed');
-        }, 200);
-        this.$emit('tapped-button', el.innerText.trim());
-      }
-    },
-    onKeyup(e) {
-      const isBackspace = e.key === 'Backspace';
-      const isEnter = e.key === 'Enter';
-      const isNumber = this.isNumber(e.key);
+<script setup>
+import { ref, onMounted } from 'vue';
 
-      if (isNumber) {
-        this.$emit('tapped-button', e.key);
-      }
+const backIcon = ref('<');
+const timeout = ref(null);
 
-      if (isBackspace) {
-        this.$emit('tapped-button', '<');
-      }
+const buttonRefs = ref({});
+const emit = defineEmits(['tapped-button']);
 
-      if (isEnter) {
-        this.$emit('tapped-button', 'Enter');
-      }
+function onTap(e) {
+  const el = e.currentTarget;
+  clearTimeout(timeout.value);
+  el.classList.add('pressed');
+  timeout.value = setTimeout(() => {
+    el.classList.remove('pressed');
+  }, 200);
+  emit('tapped-button', el.innerText.trim());
+}
 
-      if (isNumber || isBackspace || isEnter) {
-        const buttonEl = this.$refs[`button-${e.key}`];
-        if (buttonEl) {
-          buttonEl.classList.add('pressed');
+function onKeyup(e) {
+  const isBackspace = e.key === 'Backspace';
+  const isEnter = e.key === 'Enter';
+  const isNumber = isNumberKey(e.key);
+  if (isNumber) {
+    emit('tapped-button', e.key);
+  }
+  if (isBackspace) {
+    emit('tapped-button', '<');
+  }
+  if (isEnter) {
+    emit('tapped-button', 'Enter');
+  }
+  if (isNumber || isBackspace || isEnter) {
+    const buttonEl = buttonRefs.value[`button-${e.key}`];
+    buttonEl.classList.add('pressed');
+    setTimeout(() => {
+      buttonEl.classList.remove('pressed');
+    }, 200);
+  }
+}
 
-          setTimeout(() => {
-            buttonEl.classList.remove('pressed');
-          }, 200);
-        }
-      }
-    },
-    onKeyDown() {},
-    isNumber(value) {
-      const reg = /^[0-9]+$/;
-      return reg.test(value);
-    }
-  },
-};
+function onKeyDown(e) {}
+
+function isNumberKey(value) {
+  var reg = new RegExp('^[0-9]+$');
+  return reg.test(value);
+}
+
+onMounted(() => {
+  document.addEventListener('keydown', onKeyDown);
+  document.addEventListener('keyup', onKeyup);
+});
 </script>
 
-<style lang="scss" scoped>
-  @import '../styles/variables';
+<style lang="css" scoped>
+@import url("../styles/variables.css");
 
-  .button-row {
-    display: grid;
-    width: 100%;
-    grid-gap: 3px;
-    grid-template-columns: 1fr 1fr 1fr;
-    margin-bottom: 3px;
-    font-family: helvetica, sans-serif;
+.NumberPad {
+  display: flex;
+  flex-direction: column;
+}
 
-    .button {
-      height: $block-height;
-      min-height: $block-min-height;
-      max-height: $block-max-height;
-      background-color: $light-bg;
-      border-radius: 9px;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      transition: all 200ms ease;
-      backface-visibility: hidden;
-      user-select: none;
+.button-row {
+  display: grid;
+  width: 100%;
+  grid-gap: 3px;
+  grid-template-columns: 1fr 1fr 1fr;
+  margin-bottom: 3px;
+  font-family: helvetica;
+}
 
-      &:hover {
-        background-color: #ecf8ff;
-      }
+.button {
+  height: var(--block-height);
+  min-height: var(--block-min-height);
+  max-height: var(--block-max-height);
+  background-color: var(--light-bg);
+  border-radius: 9px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  transition: all 200ms ease;
+  backface-visibility: hidden;
+  user-select: none;
+}
 
-      &.pressed {
-        transform: scale(0.98);
-        opacity: 0.9;
-      }
+.button:hover {
+  background-color: #ecf8ff;
+}
 
-      &.clear {
-        color: $red;
-      }
+.button.pressed {
+  transform: scale(0.98);
+  opacity: 0.9;
+}
 
-      &.enter {
-        color: $black;
-        background-color: $red;
-        &:hover {
-          background-color: #fb7755;
-        }
-      }
-    }
-  }
+.button.clear {
+  color: var(--red);
+}
+
+.button.enter {
+  color: var(--black);
+  background-color: var(--red);
+}
+
+.button.enter:hover {
+  background-color: #fb7755;
+}
 </style>
