@@ -1,11 +1,13 @@
 <template>
   <div
+    v-show="false"
     ref="canvas"
     class="ShareCanvas"
   />
 </template>
 
-<script>
+<script setup>
+import { ref, onMounted } from 'vue';
 import Konva from 'konva';
 
 const LIGHT_BG = '#D9E4EA';
@@ -56,160 +58,165 @@ const LOGO = {
   y: GAP
 };
 
-export default {
-  props: {
-    firstPerson: {
-      type: Object,
-      default: () => ({
-        name: '',
-      }),
-    },
-    secondPerson: {
-      type: Object,
-      default: () => ({
-        name: 'person 1',
-      }),
-    },
-    expenses: {
-      type: String,
-      default: '0',
-    },
-    currency: {
-      type: String,
-      default: '$',
-    },
+const props = defineProps({
+  firstPerson: {
+    type: Object,
+    default: () => ({ name: '' }),
   },
-  mounted() {
-    const { canvas } = this.$refs;
-
-    this.stage = new Konva.Stage({
-      container: canvas,
-      width: CANVAS_WIDTH,
-      height: CANVAS_HEIGHT,
-    });
-    
-    this.layer = new Konva.Layer();
-    this.stage.add(this.layer);
-
-    const bg = new Konva.Rect({
-      x: 0,
-      y: 0,
-      width: CANVAS_WIDTH,
-      height: CANVAS_HEIGHT,
-      fill: DARK_BG,
-    });
-    this.layer.add(bg);
-
-    this.drawHeader();
-    this.drawRow1();
-    this.drawRow2();
- 
-    this.layer.draw();
+  secondPerson: {
+    type: Object,
+    default: () => ({ name: 'person 1' }),
   },
-  methods: {
-    drawHeader() {
-      const smallBox1 = new Konva.Rect({...SMALL_BOX, x: GAP, y: GAP});
-      const logo = new Konva.Rect(LOGO);
-      const smallBox2 = new Konva.Rect({...SMALL_BOX,
-        x: SMALL_BOX.width + LOGO.width + 13,
-        y: GAP
-      });
+  expenses: {
+    type: String,
+    default: '0',
+  },
+  currency: {
+    type: String,
+    default: '$',
+  },
+});
 
-      Konva.Image.fromURL('/img/logo.ea9ec231.svg', (image) => {
-        image.setAttrs({
-          x: (CANVAS_WIDTH / 2) - 20,
-          y: 35,
-          scaleX: 1,
-        });
-        this.layer.add(image);
-        this.layer.draw();
-      });
+const canvas = ref(null);
+let stage = null;
+let layer = null;
+let ROW_1_COL_1 = null;
+let ROW_1_COL_2 = null;
+let ROW_2_COL_1 = null;
+let ROW_2_COL_2 = null;
 
-      this.layer.add(smallBox1);
-      this.layer.add(logo);
-      this.layer.add(smallBox2);
-    },
-    drawRow1() {
-      this.ROW_1_COL_1 = {
-        ...MEDIUM_BOX,
-        x: 3,
-        y: HEADER_HEIGHT + 8,
-      };
+function drawHeader() {
+  const smallBox1 = new Konva.Rect({ ...SMALL_BOX, x: GAP, y: GAP });
+  const logo = new Konva.Rect(LOGO);
+  const smallBox2 = new Konva.Rect({
+    ...SMALL_BOX,
+    x: SMALL_BOX.width + LOGO.width + 13,
+    y: GAP,
+  });
 
-      this.ROW_1_COL_2 = {
-        ...LARGE_BOX,
-        x: this.ROW_1_COL_1.width + 8,
-        y: HEADER_HEIGHT + 8,
-      };
+  Konva.Image.fromURL('/img/logo.ea9ec231.svg', (image) => {
+    image.setAttrs({
+      x: CANVAS_WIDTH / 2 - 20,
+      y: 35,
+      scaleX: 1,
+    });
+    layer.add(image);
+    layer.draw();
+  });
 
-      const FIRST_NAME = {
-        ...DEFAULT_TEXT,
-        width: this.ROW_1_COL_1.width,
-        height: this.ROW_1_COL_1.height,
-        x: this.ROW_1_COL_1.x,
-        y: this.ROW_1_COL_1.y + (this.ROW_1_COL_1.height / 2) - (FONT_SIZE / 2),
-        text: this.firstPerson.name,
-      };
+  layer.add(smallBox1);
+  layer.add(logo);
+  layer.add(smallBox2);
+}
 
-      const FIRST_SALARY = {
-        ...DEFAULT_TEXT,
-        width: this.ROW_1_COL_2.width,
-        height: this.ROW_1_COL_2.height,
-        x: this.ROW_1_COL_2.x,
-        y: this.ROW_1_COL_2.y + (this.ROW_1_COL_2.height / 2) - (FONT_SIZE / 2),
-        text: `${this.currency}${this.firstPerson.salary}`,
-      };
+function drawRow1() {
+  ROW_1_COL_1 = {
+    ...MEDIUM_BOX,
+    x: 3,
+    y: HEADER_HEIGHT + 8,
+  };
 
-      this.layer.add(new Konva.Rect(this.ROW_1_COL_1));
-      this.layer.add(new Konva.Rect(this.ROW_1_COL_2));
-      this.layer.add(new Konva.Text(FIRST_NAME));
-      this.layer.add(new Konva.Text(FIRST_SALARY));
-    },
-    drawRow2() {
-      this.ROW_2_COL_1 = {
-        ...this.ROW_1_COL_1,
-        y: this.ROW_1_COL_1.y + this.ROW_1_COL_1.height + 5,
-      };
+  ROW_1_COL_2 = {
+    ...LARGE_BOX,
+    x: ROW_1_COL_1.width + 8,
+    y: HEADER_HEIGHT + 8,
+  };
 
-      this.ROW_2_COL_2 = {
-        ...this.ROW_1_COL_2,
-        y: this.ROW_1_COL_2.y + this.ROW_1_COL_2.height + 5,
-      };
+  const FIRST_NAME = {
+    ...DEFAULT_TEXT,
+    width: ROW_1_COL_1.width,
+    height: ROW_1_COL_1.height,
+    x: ROW_1_COL_1.x,
+    y: ROW_1_COL_1.y + ROW_1_COL_1.height / 2 - FONT_SIZE / 2,
+    text: props.firstPerson.name,
+  };
 
-      const SECOND_NAME = {
-        ...DEFAULT_TEXT,
-        width: this.ROW_2_COL_1.width,
-        height: this.ROW_2_COL_1.height,
-        x: this.ROW_2_COL_1.x,
-        y: this.ROW_2_COL_1.y + (this.ROW_2_COL_1.height / 2) - (FONT_SIZE / 2),
-        text: this.secondPerson.name,
-      };
+  const FIRST_SALARY = {
+    ...DEFAULT_TEXT,
+    width: ROW_1_COL_2.width,
+    height: ROW_1_COL_2.height,
+    x: ROW_1_COL_2.x,
+    y: ROW_1_COL_2.y + ROW_1_COL_2.height / 2 - FONT_SIZE / 2,
+    text: `${props.currency}${props.firstPerson.salary}`,
+  };
 
-      const SECOND_SALARY = {
-        ...DEFAULT_TEXT,
-        width: this.ROW_2_COL_2.width,
-        height: this.ROW_2_COL_2.height,
-        x: this.ROW_2_COL_2.x,
-        y: this.ROW_2_COL_2.y + (this.ROW_2_COL_2.height / 2) - (FONT_SIZE / 2),
-        text: `${this.currency}${this.secondPerson.salary}`,
-      };
+  layer.add(new Konva.Rect(ROW_1_COL_1));
+  layer.add(new Konva.Rect(ROW_1_COL_2));
+  layer.add(new Konva.Text(FIRST_NAME));
+  layer.add(new Konva.Text(FIRST_SALARY));
+}
 
-      this.layer.add(new Konva.Rect(this.ROW_2_COL_1));
-      this.layer.add(new Konva.Rect(this.ROW_2_COL_2));
-      this.layer.add(new Konva.Text(SECOND_NAME));
-      this.layer.add(new Konva.Text(SECOND_SALARY));
-    },
-    drawRow3() {
-    },
-  }
-};
+function drawRow2() {
+  ROW_2_COL_1 = {
+    ...ROW_1_COL_1,
+    y: ROW_1_COL_1.y + ROW_1_COL_1.height + 5,
+  };
+
+  ROW_2_COL_2 = {
+    ...ROW_1_COL_2,
+    y: ROW_1_COL_2.y + ROW_1_COL_2.height + 5,
+  };
+
+  const SECOND_NAME = {
+    ...DEFAULT_TEXT,
+    width: ROW_2_COL_1.width,
+    height: ROW_2_COL_1.height,
+    x: ROW_2_COL_1.x,
+    y: ROW_2_COL_1.y + ROW_2_COL_1.height / 2 - FONT_SIZE / 2,
+    text: props.secondPerson.name,
+  };
+
+  const SECOND_SALARY = {
+    ...DEFAULT_TEXT,
+    width: ROW_2_COL_2.width,
+    height: ROW_2_COL_2.height,
+    x: ROW_2_COL_2.x,
+    y: ROW_2_COL_2.y + ROW_2_COL_2.height / 2 - FONT_SIZE / 2,
+    text: `${props.currency}${props.secondPerson.salary}`,
+  };
+
+  layer.add(new Konva.Rect(ROW_2_COL_1));
+  layer.add(new Konva.Rect(ROW_2_COL_2));
+  layer.add(new Konva.Text(SECOND_NAME));
+  layer.add(new Konva.Text(SECOND_SALARY));
+}
+
+function drawRow3() {
+  // Placeholder for future implementation
+}
+
+onMounted(() => {
+  stage = new Konva.Stage({
+    container: canvas.value,
+    width: CANVAS_WIDTH,
+    height: CANVAS_HEIGHT,
+  });
+
+  layer = new Konva.Layer();
+  stage.add(layer);
+
+  const bg = new Konva.Rect({
+    x: 0,
+    y: 0,
+    width: CANVAS_WIDTH,
+    height: CANVAS_HEIGHT,
+    fill: DARK_BG,
+  });
+  layer.add(bg);
+
+  drawHeader();
+  drawRow1();
+  drawRow2();
+
+  layer.draw();
+});
 </script>
 
-<style lang="scss">
-  .ShareCanvas {
-    position: fixed;
-    left: 50%;
-    top: 50%;
-    transform: translate3d(-150%, -50%, 0);
-  }
+<style lang="css" scoped>
+.ShareCanvas {
+  position: fixed;
+  left: 50%;
+  top: 50%;
+  transform: translate3d(-150%, -50%, 0);
+}
 </style>
