@@ -1,7 +1,9 @@
-import ApplicationComponent from './Application.vue';
-import AppStore from './store/AppStore';
+import { createApp } from 'vue';
 import { createStore } from 'vuex';
-import { createApp } from 'vue' ;
+
+import Application from './Application.vue';
+import AppStore from './store/AppStore';
+import Vue2TouchEvents from 'vue2-touch-events';
 
 const store = createStore({
   modules: {
@@ -9,7 +11,7 @@ const store = createStore({
   },
 });
 
-const app = createApp(ApplicationComponent);
-
+const app = createApp(Application);
 app.use(store);
+app.use(Vue2TouchEvents);
 app.mount('#app');
